Crate quill_lexer

Converts each line of an input into a list of tokens.

Structs

Token

A single token such as an identifier or special character.

Tree

A list of tokens grouped by a set of matching brackets.

Enums

BracketType

TokenTree

A program (and by extension a line of code) is subdivided into token trees, which are essentially bracketed groups. For example, in the expression 1 + (2 + 3) + 4, the token trees are [1, +, [2, +, 3], +, 4].

TokenType

Functions

lex

Lexes a source file. This function reads a source file from disk (if not cached), splits it into tokens, and groups those tokens into token trees.